+++ src/Data/Extensible/Sum.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE PatternSynonyms #-}
 -----------------------------------------------------------------------------
 -- |
 -- Module      :  Data.Extensible.Sum
@@ -22,6 +23,7 @@ module Data.Extensible.Sum (
   , exhaust
   , picked
   , embedAssoc
+  , pattern UnionAt
   ) where
 
 import Data.Extensible.Internal
@@ -32,17 +34,20 @@ import Data.Typeable
 
 -- | The extensible sum type
 data (h :: k -> *) :| (s :: [k]) where
-  UnionAt :: !(Membership xs x) -> h x -> h :| xs
+  EmbedAt :: !(Membership xs x) -> h x -> h :| xs
 deriving instance Typeable (:|)
+
+{-# DEPRECATED UnionAt "This has renamed to EmbedAt" #-}
+pattern UnionAt a b = EmbedAt a b
 
 -- | Change the wrapper.
 hoist :: (forall x. g x -> h x) -> g :| xs -> h :| xs
-hoist f (UnionAt pos h) = UnionAt pos (f h)
+hoist f (EmbedAt p h) = EmbedAt p (f h)
 {-# INLINE hoist #-}
 
 -- | /O(1)/ lift a value.
 embed :: (x ∈ xs) => h x -> h :| xs
-embed = UnionAt membership
+embed = EmbedAt membership
 {-# INLINE embed #-}
 
 strike :: forall h x xs. (x ∈ xs) => h :| xs -> Maybe (h x)
@@ -50,16 +55,16 @@ strike = strikeAt membership
 {-# INLINE strike #-}
 
 strikeAt :: forall h x xs. Membership xs x -> h :| xs -> Maybe (h x)
-strikeAt q (UnionAt p h) = case compareMembership p q of
+strikeAt q (EmbedAt p h) = case compareMembership p q of
   Right Refl -> Just h
   _ -> Nothing
 {-# INLINE strikeAt #-}
 
 -- | /O(1)/ Naive pattern match
 (<:|) :: (h x -> r) -> (h :| xs -> r) -> h :| (x ': xs) -> r
-(<:|) r c = \(UnionAt pos h) -> runMembership pos
+(<:|) r c = \(EmbedAt pos h) -> runMembership pos
   (\Refl -> r h)
-  (\pos' -> c (UnionAt pos' h))
+  (\pos' -> c (EmbedAt pos' h))
 infixr 1 <:|
 {-# INLINE (<:|) #-}
 
@@ -69,11 +74,11 @@ exhaust _ = error "Impossible"
 
 -- | A traversal that tries to point a specific element.
 picked :: forall f h x xs. (x ∈ xs, Applicative f) => (h x -> f (h x)) -> h :| xs -> f (h :| xs)
-picked f u@(UnionAt pos h) = case compareMembership (membership :: Membership xs x) pos of
-  Right Refl -> fmap (UnionAt pos) (f h)
+picked f u@(EmbedAt pos h) = case compareMembership (membership :: Membership xs x) pos of
+  Right Refl -> fmap (EmbedAt pos) (f h)
   _ -> pure u
 {-# INLINE picked #-}
 
 embedAssoc :: Associate k a xs => h (k ':> a) -> h :| xs
-embedAssoc = UnionAt association
+embedAssoc = EmbedAt association
 {-# INLINE embedAssoc #-}